move named_semaphore to a header
